[Commands] Make `migrate` a sub-command of `swift package`
