+++ Sources/_InternalTestSupport/SwiftPMProduct.swift
@@ -23,7 +23,6 @@ import enum TSCBasic.ProcessEnv
 /// helper method to execute them.
 public enum SwiftPM {
     case Build
-    case Migrate
     case Package
     case Registry
     case Test
@@ -38,8 +37,6 @@ extension SwiftPM {
         switch self {
         case .Build:
             return "swift-build"
-        case .Migrate:
-            return "swift-migrate"
         case .Package:
             return "swift-package"
         case .Registry:
