--- Sources/swift-migrate/CMakeLists.txt
@@ -1,18 +1,0 @@
-# This source file is part of the Swift open source project
-#
-# Copyright (c) 2025 Apple Inc. and the Swift project authors
-# Licensed under Apache License v2.0 with Runtime Library Exception
-#
-# See http://swift.org/LICENSE.txt for license information
-# See http://swift.org/CONTRIBUTORS.txt for Swift project authors
-
-add_executable(swift-migrate
-    Entrypoint.swift)
-target_link_libraries(swift-build PRIVATE
-    Commands)
-
-target_compile_options(swift-migrate PRIVATE
-  -parse-as-library)
-
-install(TARGETS swift-migrate
-  DESTINATION bin)
